legacy/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==17.0.0
boto3==1.42.69
botocore==1.42.69
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
distro==1.9.0
filelock==3.25.2
fsspec==2026.3.0
google-auth==2.49.1
google-genai==1.67.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pillow==12.2.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2026.3.32
requests==2.32.5
rich==14.3.3
s3transfer==0.16.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yt-dlp==2026.3.13
